% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.growthcurve.R
\name{as.data.frame.growthcurve}
\alias{as.data.frame.growthcurve}
\title{Create a data frame from a fit}
\usage{
\method{as.data.frame}{growthcurve}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{A fit for some growth data (a \code{growthcurve} object)}

\item{row.names}{\code{NULL} or a character vector giving the row names for
the data frame. Missing values are not allowed. (ignored)}

\item{optional}{(ignored)}

\item{...}{Additional arguments (ignored)}
}
\value{
A data.frame containing the parameters for the fit
}
\description{
\code{as.data.frame} creates a data frame that contains information about
the fit
}

