% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.growthcurve.R
\name{augment.growthcurve}
\alias{augment.growthcurve}
\title{Augment Growth Data According to a Tidied Model}
\usage{
augment.growthcurve(x, ...)
}
\arguments{
\item{x}{A fit for some growth data (a \code{growthcurve} object)}

\item{...}{Additional arguments (not used)}
}
\value{
an expanded data frame
}
\description{
Add columns to the growth data set such as predictions and residuals.
Currently, this just wraps the \code{augment} function for the type of model
used, so models fit with \code{\link{fit_growth_logistic}} will differ from
\code{\link{fit_growth_linear}}, \code{\link{fit_growth_loess}}, etc.
}
\note{
This function requires the \pkg{broom} package.
}
\examples{
\dontrun{
myfit <- fit_growth_gompertz(mydata, Time, OD600)
augment(myfit)}

}

