% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.gcfit.R
\name{autoplot.gcfit}
\alias{autoplot.gcfit}
\title{Create a ggplot for a growth curve}
\usage{
autoplot.gcfit(object, show_data = TRUE, show_maxrate = TRUE,
  show_asymptote = FALSE, show_lag = FALSE, xtrans = "identity",
  ytrans = "log10", xlab = NULL, ylab = NULL, title = NULL,
  subtitle = NULL, caption = NULL)
}
\arguments{
\item{object}{A fit for some growth data}

\item{show_data}{Whether or not to show original data points (default: \code{TRUE})}

\item{show_maxrate}{Whether or not to indicate the maximum growth rate (default: \code{TRUE})}

\item{show_asymptote}{Whether or not to show the asymptote (default: \code{TRUE})}

\item{show_lag}{Whether or not to indicate where lag phase ends (default \code{FALSE}}
}
\value{
A ggplot object
}
\description{
Create a ggplot for a growth curve
}
\examples{
\dontrun{
# Get a logistic fit for some data and extract its residuals
lfit <- fit_growth_logistic(df=mydata, Time, OD600)
autoplot(lfit, title="My Growth Data")}

}

