% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_gfgompertz.exp.R
\name{fit_growth_gfgompertz.exp}
\alias{fit_growth_gfgompertz.exp}
\alias{fit_growth_gfgompertz.exp_}
\title{Fit a Modified Gompertz Growth Curve to Growth Data (using grofit)}
\usage{
fit_growth_gfgompertz.exp(df, time, data, ...)

fit_growth_gfgompertz.exp_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link{gcFitModel}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{growthcurve} object with the following fields:
\itemize{
    \item \code{type}: String describing the type of fit (here, "grofit/gompertz.exp")
    \item \code{parameters}: Parameters for the fitted model. A list with
    fields:
    \itemize{
        \item{TODO}: TODO
    }
    \item \code{model}: An \code{\link{nls}} object containing the fit.
    \item \code{data}: A list containing the input data frame (\code{df}),
      the name of the column containing times (\code{time_col}), and the
      name of the column containing growth values (\code{data_col}).
    \item \code{grofit}: An object of class \code{gcFitModel}
}
}
\description{
\code{fit_growth_gfgompertz.exp} fits a modified Gompertz curve to a tidy
data set using \code{\link[grofit]{gcFitModel}} from the \pkg{grofit}
package.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_gfgompertz(mydata, Time, OD600)}
}
\seealso{
\code{\link{fit_growth_gompertz}}, growthcurve's native function
for fitting Gompertz curves
}

