% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_gfspline.R
\name{fit_growth_gfspline}
\alias{fit_growth_gfspline}
\alias{fit_growth_gfspline_}
\title{Fit Smooth Splines to Growth Data (Using grofit)}
\usage{
fit_growth_gfspline(df, time, data, ...)

fit_growth_gfspline_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link[grofit]{gcFitSpline}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{growthcurve} object with the following fields:
\itemize{
    \item \code{type}: String describing the type of fit (here, "grofit/spline")
    \item \code{parameters}: Growth parameters from the fitted model. A list
    with fields:
    \itemize{
        \item{TODO}: TODO
    }
    \item \code{model}: An \code{\link{smooth.spline}} object containing the "fit".
    \item \code{data}: A list containing the input data frame (\code{df}),
      the name of the column containing times (\code{time_col}), and the
      name of the column containing growth values (\code{data_col}).
    \item \code{grofit}: An object of class \code{gcFitSpline}
}
}
\description{
\code{fit_growth_gfspline} fits a smooth spline to a tidy growth data set
using \code{\link[grofit]{gcFitSpline}} from the \pkg{grofit} package.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_gfspline(mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_gfspline_(df=mydata, time_col = "Time", data_col = "OD600")}

}
\seealso{
\code{\link{fit_growth_spline}}, growthcurve's native function    
for fitting smooth splines to growth data
}

