% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_gompertz.R
\name{fit_growth_gompertz}
\alias{fit_growth_gompertz}
\alias{fit_growth_gompertz_}
\title{Fit a Gompertz Curve to Growth Data}
\usage{
fit_growth_gompertz(df, time, data, ...)

fit_growth_gompertz_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{a data frame}

\item{time}{column in \code{df} that contains time data}

\item{data}{column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments to \code{\link[stats]{nls}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{\link{growthcurve}} object
}
\description{
\code{fit_growth_gompertz} fits a Gompertz curve to a tidy growth data set
using nonlinear least squares
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_gompertz(mydata, Time, OD600)}

\dontrun{
fit_growth_gompertz_(mydata, "Time", "OD600")
}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Gompertz_function}
}

