% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_grofit_spline.R
\name{fit_growth_grofit_spline}
\alias{fit_growth_grofit_spline}
\alias{fit_growth_grofit_spline_}
\title{Fit Smooth Splines to Growth Data (Using grofit)}
\usage{
fit_growth_grofit_spline(df, time, data, ...)

fit_growth_grofit_spline_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{a data frame}

\item{time}{column in \code{df} that contains time data}

\item{data}{column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link[grofit]{gcFitSpline}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{growthcurve} object
}
\description{
\code{fit_growth_grofit_spline} fits a smooth spline to a tidy growth data set
using \code{\link[grofit]{gcFitSpline}} from the \pkg{grofit} package.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_grofit_spline(mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_grofit_spline_(df=mydata, time_col = "Time", data_col = "OD600")}

}
\seealso{
\code{\link{fit_growth_spline}}, growthcurve's native function    
for fitting smooth splines to growth data

\link{grofit_wrappers}
}

