% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_linear.R
\name{fit_growth_linear}
\alias{fit_growth_linear}
\alias{fit_growth_linear_}
\title{Fit a Linear Model  to Growth Data}
\usage{
fit_growth_linear(df, time, data, ...)

fit_growth_linear_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments to \code{\link{nls}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{\link{growthcurve}} object with the following fields:
\itemize{
    \item \code{type}: The type of fit (here "linear")
    \item \code{parameters}: Growth parameters from the fitted model. A list
    with fields:
    \itemize{
        \item{TODO}: TODO
    }
    \item \code{model}: An \code{\link{nls}} object containing the fit.
    \item \code{data}: A list containing the input data frame (\code{df}),
      the name of the column containing times (\code{time_col}), and the
      name of the column containing growth values (\code{data_col}).
}
}
\description{
\code{fit_growth_linear} fits a linear model to a tidy growth data set
using nonlinear least squares
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_linear(mydata, Time, OD600)}

\dontrun{
fit_growth_linear_(mydata, "Time", "OD600")
}
}

