% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_loess.R
\name{fit_growth_loess}
\alias{fit_growth_loess}
\alias{fit_growth_loess_}
\title{Fit a Curve to Growth Data using LOESS}
\usage{
fit_growth_loess(df, time, data, ...)

fit_growth_loess_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{a data frame}

\item{time}{column in \code{df} that contains time data}

\item{data}{column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments to \code{\link[stats]{loess}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{\link{growthcurve}} object
}
\description{
\code{fit_growth_loess} fits curve to a tidy growth data set using LOESS
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_loess(mydata, Time, OD600)}

\dontrun{
fit_growth_loess_(mydata, "Time", "OD600")
}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Local_regression}
}

