% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_growth_parametric.R
\name{fit_growth_parametric}
\alias{fit_growth_parametric}
\alias{fit_growth_parametric_}
\title{Fit parametric models to growth data}
\usage{
fit_growth_parametric(df, time, growth, ...)

fit_growth_parametric_(df, time_col, growth_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{growth}{Name of the column in \code{df} that contains growth data}

\item{...}{Additional arguments for \code{\link{gcFitModel}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{growth_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A list of types \code{gcfit} and \code{gcFitModel}
    \item{raw.time}{Raw data given to the function; equivalent to \code{time}}
    \item{raw.data}{Raw data given to the function; equivalent to \code{data}}
    \item{gcID}{Vector (of any length) identifying the growth curve data}
    \item{fit.time}{Vector of fitted concentration values}
    \item{fit.data}{Vector of fitted growth values}
    \item{parameters}{List of estimated growth values}
    \item{A}{Maximum growth value}
    \item{mu}{Maximum slope}
    \item{lambda}{Lag phase}
    \item{integral}{Integral under growth curve}
    \item{model}{String naming the parametric model used}
    \item{nls}{\code{nls} object generated by the R internal function \code{\link{nls}}}
    \item{reliable}{Logical, indicating wether the provided data is reliable (to be set manually)}
    \item{fitFlag}{Logical, indicating wether a model could be successfully fitted to data}
    \item{control}{Object of class \code{grofit.control} containing list of options passed to the function as \code{control}}
    \item{fit_type}{The type of fit, here 'model'}
}
\description{
\code{fit_growth_parametric} uses \code{\link{gcFitModel}} from the
\pkg{grofit} package to fit a parametric model to growth data. Several
candidate models are fitted, and the model with the best AIC is returned.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_parametric(df=mydata, Time, OD600)}
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_parametric_(df=mydata, time_col='Time', growth_col='OD600')}
}
\seealso{
For specific parametric models, see \code{\link{fit_growth_logistic}}, \code{\link{fit_growth_richards}}, \code{\link{fit_growth_gompertz}}, \code{\link{fit_growth_gompertz.exp}}

\code{\link{gcFitModel}}
}

