% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_richards.R
\name{fit_growth_richards}
\alias{fit_growth_richards}
\alias{fit_growth_richards_}
\title{Fit a Richards growth curve to the given data}
\usage{
fit_growth_richards(df, time, data, ...)

fit_growth_richards_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link{gcFitModel}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\description{
Fit a Richards growth curve to the given data
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_richards(mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_richards_(df=mydata, time_col='Time', data_col='OD600')}

}
\seealso{
\code{\link{fit_growth_parametric}}

\code{\link{richards}}

\url{https://en.wikipedia.org/wiki/Generalised_logistic_function}
}

