% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_spline.R
\name{fit_growth_spline}
\alias{fit_growth_spline}
\alias{fit_growth_spline_}
\title{Fit smooth splines to growth data}
\usage{
fit_growth_spline(df, time, data, ...)

fit_growth_spline_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link{gcFitSpline}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A list of types \code{gcfit} and \code{gcfitspline} that contains:
\item{\code{type}}{The type of model (always "spline")}
\item{\code{model}}{The model used (e.g., "spline"). Models fit by grofit are prefixed with \code{grofit::}}
\item{\code{grofit}}{For models fit with grofit, the results from \code{\link{gcFitSpline}}}
\item{\code{parameters}}{Model parameters}
\item{\code{fit}}{Data fitted to the model}
\item{\code{raw}}{The original data frame and the names of columns used}
}
\description{
\code{fit_growth_spline} uses \code{\link{gcFitSpline}} from the \pkg{grofit}
package to fit a smoothed spline to growth data.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_spline(mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_spline_(df=mydata, time_col='Time', data_col='OD600')}

}

