% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_functions.R
\name{gc_save_report_from_builders}
\alias{gc_save_report_from_builders}
\title{Saves plots in single report from plot builders (batch only)}
\usage{
gc_save_report_from_builders(plot_builders, file, plate_name = NULL)
}
\description{
Saves plots in single report from plot builders (batch only)
}
