% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_functions.R
\name{gc_write_summaries}
\alias{gc_write_summaries}
\title{Saves summary files}
\usage{
gc_write_summaries(
  core,
  params,
  instrument,
  out_dir,
  region,
  raw_data_format = NULL,
  design_file_format = NULL
)
}
\description{
Saves summary files
}
