% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\name{growthcurve}
\alias{growthcurve}
\alias{is.growthcurve}
\title{Create \code{growthcurve} Objects}
\usage{
growthcurve(type, model, fit, f, parameters, df, time_col, data_col)

is.growthcurve(x)
}
\arguments{
\item{type}{A string describing the fit (e.g., "logistic")}

\item{model}{A model object, such as a \code{nls} or \code{smooth.spline}}

\item{fit}{A list containing \code{x} and \code{y} values for the fit}

\item{f}{A function \code{f(x)} that maps a time value \code{x} to a growth
value using the model}

\item{parameters}{A list containing growth parameters (TODO details)}

\item{df}{A data frame containing the raw growth data}

\item{time_col}{A string with the name of the column containing time data}

\item{data_col}{A string with the name of the column containing the growth
data}

\item{x}{A potential \code{growthcurve} object}
}
\value{
A \code{growthcurve} object

\code{is.growthcurve} returns a logical value indicating whether
\code{x} is a \code{growthcurve} object (\code{TRUE}) or not (\code{FALSE}).
}
\description{
\code{growthcurve} creates \code{growthcurve} objects, which
contain information about a function fitted to biological growth data

\code{is.growthcurve} determines whether or not the given object
is a growthcurve
}

