% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\docType{data}
\name{kvr}
\alias{kvr}
\title{Bacterial growth in a microtiter plate}
\format{A data frame with 115320 rows and 5 variables}
\usage{
data(kvr)
}
\description{
\describe{
    \item{Time}{The time at which the reading was taken (seconds)}
    \item{Well}{The microtiter plate well in which the population was grown (A1-H12)}
    \item{OD600}{The density of the population, measured at 600 nm (absorbance units)}
    \item{Strain}{The strain of bacteria. \code{control} indicates a well without bacteria.}
    \item{Environment}{The environment in which cells were cultured}
}
}
\author{
Katrina van Raay (data)

Brian Connelly (formatting)
}

