% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.gcfit.R
\name{plot.gcfit}
\alias{plot.gcfit}
\title{Plot growth data and its fitted growth curve}
\usage{
\method{plot}{gcfit}(object, ...)
}
\arguments{
\item{object}{A fit for some growth data}

\item{...}{Optional arguments (not currently used)}
}
\description{
Plot growth data and its fitted growth curve
}
\examples{
\dontrun{
# Get a logistic fit for some data and extract its residuals
lfit <- fit_growth_logistic(df=mydata, Time, OD600)
plot(lfit)}
}

