% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.growthcurve.R
\name{plot.growthcurve}
\alias{plot.growthcurve}
\title{Plot Growth Data and its Fitted Growth Curve}
\usage{
\method{plot}{growthcurve}(x, y = NULL, show_fit = TRUE, show_data = TRUE,
  show_maxrate = TRUE, show_asymptote = FALSE, ...)
}
\arguments{
\item{x}{A fit for some growth data (a \code{growthcurve} object)}

\item{y}{Not used}

\item{show_fit}{Whether or not to show the fitted curve (default: \code{TRUE})}

\item{show_data}{Whether or not to show the original data (default \code{TRUE})}

\item{show_maxrate}{Whether or not to show a tangent line where the maximum growth rate occurs (default \code{TRUE})}

\item{show_asymptote}{Whether or not to indicate the maximum growth level (default \code{FALSE})}

\item{...}{Optional formatting arguments. Includes \code{xlab}, \code{ylab},
\code{title}, \code{subtitle},
\code{fit.color}, \code{fit.linetype}, \code{fit.size},
\code{data.color}, \code{data.fill}, \code{data.shape},
\code{data.size}, \code{data.stroke},
\code{maxrate.color}, \code{maxrate.linetype}, \code{maxrate.size},
\code{asymptote.color}, \code{asymptote.linetype},
\code{asymptote.size}}
}
\description{
Plot Growth Data and its Fitted Growth Curve
}
\examples{
\dontrun{
# Get a logistic fit for some data and plot it
lfit <- fit_growth_logistic(mydata, Time, OD600)
plot(lfit)
}

}

