% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.growthcurve.R
\name{predict.growthcurve}
\alias{predict.growthcurve}
\title{Predict method for Growth Curves}
\usage{
\method{predict}{growthcurve}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class inheriting from "growthcurve"}

\item{newdata}{An optional vector of values with which to predict. If
omitted, the fitted values are used.}

\item{...}{Additional arguments (not used)}
}
\value{
A numeric vector of predictions
}
\description{
Predict method for Growth Curves
}
\examples{
\dontrun{
gc1 <- fit_growth_logistic(mydata, Time, OD600)
predict(gc1)
}
}

