% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\docType{data}
\name{pseudomonas}
\alias{pseudomonas}
\title{Growth data for Pseudomonas aeruginosa strains}
\format{A data frame with 180 rows and 4 variables}
\usage{
data(pseudomonas)
}
\description{
A dataset containing growth information for two strains of the bacterium
\emph{Pseudomonas aeruginosa} grown in a limiting medium for 48 hours.
Growth was estimated by counting colonies on LB agar plates.
}
\details{
\describe{
    \item{Time}{The time at which growth was measured (hours)}
    \item{Replicate}{The replicate population (unitless)}
    \item{Strain}{The strain, either PAO1 (Wild type) or Mutant}
    \item{CFUmL}{The current yield (colony forming units per mL)}
}
}
\keyword{datasets}

