% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.growthcurve.R
\name{residuals.growthcurve}
\alias{residuals.growthcurve}
\title{Get model residuals from fits}
\usage{
\method{residuals}{growthcurve}(object, ...)
}
\arguments{
\item{object}{A fit for some growth data (a \code{growthcurve} object)}

\item{...}{Additional arguments (not used)}
}
\value{
A numeric list containing the fit's residuals
}
\description{
residuals gives the residuals, or the difference between the observed growth
data and the fitted values, for a given fit.
}
\examples{
\dontrun{
# Get a logistic fit for some data and extract its residuals
lfit <- fit_growth_logistic(mydata, Time, OD600)
residuals(lfit)}

}

