% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_functions.R
\name{run_gc}
\alias{run_gc}
\title{Run growth curve analysis}
\usage{
run_gc(
  rawdatafile,
  designfile,
  design_vars = NULL,
  hrs,
  interval,
  minod,
  maxod,
  instrument = "plate_reader",
  prefix = "",
  blank_mode = "plate",
  batch = TRUE,
  cancel = NULL,
  region = "US",
  raw_data_format = NULL,
  design_file_format = NULL,
  debug_logfile = NULL
)
}
\description{
Run growth curve analysis
}
