% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.gcfit.R
\name{tidy.gcfit}
\alias{tidy.gcfit}
\title{Tidy a fitted growth curve into a summary data.frame}
\usage{
\method{tidy}{gcfit}(x, ...)
}
\arguments{
\item{x}{A fit for some growth data}

\item{...}{}
}
\value{
A data.frame
}
\description{
\code{tidy.gcfit} is intended to provide a summary data frame for a growth
curve that is compatible with- and follows the naming conventions of
\code{\link{broom}}.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
myfit <- fit_growth_gompertz(mydata, Time, OD600)
tidy(myfit)}

}

