% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.growthcurve.R
\name{tidy.growthcurve}
\alias{tidy.growthcurve}
\title{Tidy a fitted growth curve into a summary data.frame}
\usage{
tidy.growthcurve(x, ...)
}
\arguments{
\item{x}{A fit for some growth data (a \code{growthcurve} object)}

\item{...}{Additional arguments (not used)}
}
\value{
A data.frame
}
\description{
\code{tidy.growthcurve} provides a summary data frame for a growth
curve that is compatible with- and follows the naming conventions of
\code{\link[broom]{broom-package}}.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
myfit <- fit_growth_gompertz(mydata, Time, OD600)
tidy(myfit)}

}

