% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.data.frame.gcfit.R
\name{as.data.frame.gcfit}
\alias{as.data.frame.gcfit}
\title{Create a data frame from a fit}
\usage{
\method{as.data.frame}{gcfit}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A fit for some growth data}

\item{row.names}{\code{NULL} or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{(ignored)}

\item{...}{Additional arguments (ignored)}
}
\value{
A data frame containing the time and growth values for the fit
}
\description{
\code{as.data.frame} creates a data frame that contains data tracing the
fitted curve. The column names are the same as those used when the fit was
created.
}
\examples{
\dontrun{
# Plot some sample data and overlay a fit line
library(ggplot2)
library(growthcurve)

sampledata <- data.frame(Time=1:30, OD600=1/(1+exp(0.5*(15-1:30)))+rnorm(30)/20)
lfit <- fit_growth_logistic(sampledata, Time, OD600)

ggplot(sampledata, aes(x=Time, y=OD600)) +
    geom_point(shape=1) +
    geom_line(data=as.data.frame(lfit))}
}

