% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.R
\name{fit_growth}
\alias{fit_growth}
\alias{fit_growth_}
\title{Fit a growth curve to the given data}
\usage{
fit_growth(df, time, data, type = "parametric", ...)

fit_growth_(df, time_col, data_col, type = "parametric", ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{type}{Type of model to fit. One of \code{parametric} (default),
\code{gompertz}, \code{gompertz.exp}, \code{logistic}, \code{parametric},
\code{richards}, or \code{spline}. If \code{parametric} is chosen, several       
candidate models are fitted, and the model with the best AIC is returned.}

\item{...}{Additional arguments for \code{\link{gcFitModel}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A list of types \code{gcfit} and \code{gcfitparametric} that contains:
\item{\code{type}}{The type of model (always "parametric")}
\item{\code{model}}{The model used (e.g., "logistic"). Models fit by grofit are prefixed with \code{grofit::}}
\item{\code{grofit}}{For models fit with grofit, the results from \code{\link[grofit]{gcFitModel}}}
\item{\code{parameters}}{Model parameters}
\item{\code{fit}}{Data fitted to the model}
\item{\code{data}}{The original data frame and the names of columns used}
}
\description{
\code{fit_growth} fits a growth curve using either a parametric model or 
splines.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth(mydata, Time, OD600, type = "logistic")}

}
\seealso{
\code{\link{fit_growth_gfparametric}}

\code{\link{fit_growth_spline}}

\url{https://en.wikipedia.org/wiki/Generalised_logistic_function}
}

