% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.R
\name{fit_growth}
\alias{fit_growth}
\alias{fit_growth_}
\title{Fit a growth curve to the given data}
\usage{
fit_growth(df, time, data, type = "parametric", ...)

fit_growth_(df, time_col, data_col, type = "parametric", ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data}

\item{type}{Type of model to fit. One of \code{gompertz}, \code{gompertz.exp}, \code{logistic}, \code{parametric}, \code{richards}, \code{spline} (default: \code{parametric})}

\item{...}{Additional arguments for \code{\link{gcFitModel}}}
}
\description{
Fit a growth curve to the given data
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth(df=mydata, Time, OD600, type = "logistic")}

}
\seealso{
\code{\link{fit_growth_parametric}}

\code{\link{fit_growth_spline}}

\url{https://en.wikipedia.org/wiki/Generalised_logistic_function}
}

