% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.R
\name{fit_growth}
\alias{fit_growth}
\alias{fit_growth_}
\title{Fit a Growth Curve to the Given Data}
\usage{
fit_growth(df, time, data, model = "logistic", ...)

fit_growth_(df, time_col, data_col, model = "logistic", ...)
}
\arguments{
\item{df}{a data frame}

\item{time}{column in \code{df} that contains time data}

\item{data}{column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{model}{name of the model to fit. One of:
\itemize{
    \item \code{logistic}: a logistic function (default)
    \item \code{logistic4p}: a 4-parameter logistic function
    \item \code{gompertz}: a Gompertz function
    \item \code{loess}: a LOESS curve
    \item \code{linear}: a linear function
    \item \code{spline}: spline interpolation of the data
    \item \code{grofit_parametric}: a parametric function. \pkg{grofit} will attempt to fit each of the following models and return the best fit according to AIC.
    \item \code{grofit_logistic}: a logistic function (using \pkg{grofit})
    \item \code{grofit_gompertz}: a Gompertz function (using \pkg{grofit})
    \item \code{grofit_gompertz.exp}: a modified Gompertz function (using \pkg{grofit})
    \item \code{grofit_richards}: a Richards growth curve (using \pkg{grofit})
    \item \code{grofit_spline}: spline interpolation (using \pkg{grofit})
}}

\item{...}{Additional arguments for the specific model function}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{\link{growthcurve}} object
}
\description{
\code{fit_growth} fits a growth curve using either a parametric model or 
splines.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth(mydata, Time, OD600, model = "logistic")}

}

