% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.R
\name{fit_growth}
\alias{fit_growth}
\alias{fit_growth_}
\title{Fit a growth curve to the given data}
\usage{
fit_growth(df, time, data, model = "logistic", ...)

fit_growth_(df, time_col, data_col, model = "logistic", ...)
}
\arguments{
\item{df}{a data frame}

\item{time}{column in \code{df} that contains time data}

\item{data}{column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{model}{name of the model to fit. One of:
\itemize{
    \item \code{logistic}: TODO (default)
    \item \code{logistic4p}: TODO
    \item \code{gompertz}: TODO
    \item \code{loess}: TODO
    \item \code{linear}: TODO
    \item \code{spline}: TODO
    \item \code{TODO}: TODO
}}

\item{...}{Additional arguments for the specific model function}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
TODO
}
\description{
\code{fit_growth} fits a growth curve using either a parametric model or 
splines.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth(mydata, Time, OD600, model = "logistic")}

}

