% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_grofit_parametric.R
\name{fit_growth_grofit_parametric}
\alias{fit_growth_grofit_parametric}
\alias{fit_growth_grofit_parametric_}
\title{Fit a Parametric Model to Growth Data (using grofit)}
\usage{
fit_growth_grofit_parametric(df, time, data, ...)

fit_growth_grofit_parametric_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link{gcFitModel}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{growthcurve} object with the following fields:
\itemize{
    \item \code{type}: String describing the type of fit
    \item \code{parameters}: Growth parameters from the fitted model. A list
    with fields:
    \itemize{
        \item{TODO}: TODO
    }
    \item \code{model}: An \code{\link{nls}} object containing the fit.
    \item \code{data}: A list containing the input data frame (\code{df}),
      the name of the column containing times (\code{time_col}), and the
      name of the column containing growth values (\code{data_col}).
    \item \code{grofit}: An object of class \code{gcFitModel}
}
}
\description{
\code{fit_growth_grofit_parametric} allows you to fit a parametric model to a
tidy growth data set using \code{\link[grofit]{gcFitModel}} from the
\pkg{grofit} package. Several candidate models are fitted, and the model with
the best AIC is returned.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_grofit_parametric(mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_grofit_parametric_(mydata, time_col = "Time", data_col = "OD600")}

}
\seealso{
For specific parametric models, see \code{\link{fit_growth_grofit_logistic}}, \code{\link{fit_growth_grofit_gompertz}}, \code{\link{fit_growth_grofit_gompertz.exp}}, \code{\link{fit_growth_grofit_richards}}
}

