% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_grofit_parametric.R
\name{fit_growth_grofit_parametric}
\alias{fit_growth_grofit_gompertz}
\alias{fit_growth_grofit_gompertz.exp}
\alias{fit_growth_grofit_gompertz.exp_}
\alias{fit_growth_grofit_gompertz_}
\alias{fit_growth_grofit_logistic}
\alias{fit_growth_grofit_logistic_}
\alias{fit_growth_grofit_parametric}
\alias{fit_growth_grofit_parametric_}
\alias{fit_growth_grofit_richards}
\alias{fit_growth_grofit_richards_}
\title{Fit a Parametric Model to Growth Data (using grofit)}
\usage{
fit_growth_grofit_parametric(df, time, data, ...)

fit_growth_grofit_logistic(df, time, data, ...)

fit_growth_grofit_gompertz(df, time, data, ...)

fit_growth_grofit_gompertz.exp(df, time, data, ...)

fit_growth_grofit_richards(df, time, data, ...)

fit_growth_grofit_parametric_(df, time_col, data_col, ...)

fit_growth_grofit_logistic_(df, time_col, data_col, ...)

fit_growth_grofit_gompertz_(df, time_col, data_col, ...)

fit_growth_grofit_gompertz.exp_(df, time_col, data_col, ...)

fit_growth_grofit_richards_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{a data frame}

\item{time}{column in \code{df} that contains time data}

\item{data}{column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link[grofit]{gcFitModel}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
TODO
}
\description{
\code{fit_growth_grofit_parametric} allows you to fit a parametric model to a
tidy growth data set using \code{\link[grofit]{gcFitModel}} from the
\pkg{grofit} package. Several candidate models are fitted, and the model with
the best AIC is returned.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_grofit_parametric(mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_grofit_parametric_(mydata, time_col = "Time", data_col = "OD600")}

}

