% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_loess.R
\name{fit_growth_loess}
\alias{fit_growth_loess}
\alias{fit_growth_loess_}
\title{Fit a Curve to Growth Data using LOESS}
\usage{
fit_growth_loess(df, time, data, ...)

fit_growth_loess_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{a data frame}

\item{time}{column in \code{df} that contains time data}

\item{data}{column in \code{df} that contains growth data
(default: \code{TRUE})}

\item{...}{Additional arguments to \code{\link{loess}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A \code{\link{growthcurve}} object with the following fields:
\itemize{
    \item \code{type}: The type of fit (here "loess")
    \item \code{parameters}: Growth parameters from the fitted model. A list
    with fields:
    \itemize{
        \item{TODO}: TODO
    }
    \item \code{model}: An \code{\link{nls}} object containing the fit.
    \item \code{data}: A list containing the input data frame (\code{df}),
      the name of the column containing times (\code{time_col}), and the
      name of the column containing growth values (\code{data_col}).
}
}
\description{
\code{fit_growth_loess} fits curve to a tidy growth data set using LOESS
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_loess(mydata, Time, OD600)}

\dontrun{
fit_growth_loess_(mydata, "Time", "OD600")
}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Local_regression}
}

