% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_logistic.R
\name{fit_growth_logistic}
\alias{fit_growth_logistic}
\alias{fit_growth_logistic_}
\title{Title TODO}
\usage{
fit_growth_logistic(df, time, data, ...)

fit_growth_logistic_(df, time_col, data_col, ...)
}
\arguments{
\item{df}{TODO}

\item{time}{TODO}

\item{data}{TODO}

\item{...}{TODO}

\item{time_col}{TODO}

\item{data_col}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
\dontrun{
TODO
}
}

