% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_parametric.R
\name{fit_growth_parametric}
\alias{fit_growth_parametric}
\alias{fit_growth_parametric_}
\title{Fit parametric models to growth data}
\usage{
fit_growth_parametric(df, time, data, include_grofit = TRUE, ...)

fit_growth_parametric_(df, time_col, data_col, include_grofit = TRUE, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data}

\item{include_grofit}{Whether or not to include result object from grofit
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link{gcFitModel}}}
}
\value{
A list of type \code{gcfit} that contains (TODO - this is under construction). Use \code{names()} to see these items.
}
\description{
\code{fit_growth_parametric} uses \code{\link{gcFitModel}} from the
\pkg{grofit} package to fit a parametric model to growth data. Several
candidate models are fitted, and the model with the best AIC is returned.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_parametric(df=mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_parametric_(df=mydata, time_col='Time', data_col='OD600')}

}
\seealso{
For specific parametric models, see \code{\link{fit_growth_logistic}}, \code{\link{fit_growth_gompertz}}, \code{\link{fit_growth_gompertz.exp}}, \code{\link{fit_growth_richards}}
}

