% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_spline.R
\name{fit_growth_spline}
\alias{fit_growth_spline}
\alias{fit_growth_spline_}
\title{Fit smooth splines to growth data}
\usage{
fit_growth_spline(df, time, data, include_grofit = TRUE, ...)

fit_growth_spline_(df, time_col, data_col, include_grofit = TRUE, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{data}{Name of the column in \code{df} that contains growth data}

\item{include_grofit}{Whether or not to include result object from grofit
(default: \code{TRUE})}

\item{...}{Additional arguments for \code{\link{gcFitSpline}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{data_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A list of type \code{gcfit} that contains (TODO - this is under construction). Use \code{names()} to see these items.
}
\description{
\code{fit_growth_spline} uses \code{\link{gcFitSpline}} from the \pkg{grofit}
package to fit a smoothed spline to growth data.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_spline(df=mydata, Time, OD600)}

\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_spline_(df=mydata, time_col='Time', data_col='OD600')}

}
\seealso{
\code{\link{gcFitSpline}}
}

