% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_growth_spline.R
\name{fit_growth_spline}
\alias{fit_growth_spline}
\alias{fit_growth_spline_}
\title{Fit smooth splines to growth data}
\usage{
fit_growth_spline(df, time, growth, ...)

fit_growth_spline_(df, time_col, growth_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{time}{Name of the column in \code{df} that contains time data}

\item{growth}{Name of the column in \code{df} that contains growth data}

\item{...}{Additional arguments for \code{\link{gcFitSpline}}}

\item{time_col}{String giving the name of the column in \code{df} that
contains time data}

\item{growth_col}{String giving the name of the column in \code{df} that
contains growth data}
}
\value{
A list of types \code{gcfit} and \code{gcFitSpline}
    \item{raw.time}{Raw data given to the function; equivalent to \code{time}}
    \item{raw.data}{Raw data given to the function; equivalent to \code{data}}
    \item{gcID}{Vector (of any length) identifying the growth curve data}
    \item{fit.time}{Vector of fitted concentration values}
    \item{fit.data}{Vector of fitted growth values}
    \item{parameters}{List of estimated growth values from spline fit: \code{A}, maximum growth value;
\code{mu}, maximum slope; \code{lambda}, lag-phase, \code{integral},
integral under growth curve}
    \item{parametersLowess}{List of estimated growth values from \code{\link{lowess}} fit: \code{A}, maximum growth value; \code{mu}, maximum slope; \code{lambda}, lag-phase}
    \item{spline}{\code{smooth.spline} object generated by \code{\link{smooth.spline}}}
    \item{reliable}{Logical, indicating wether the provided data is reliable (to be set manually)}
    \item{fitFlag}{Logical, indicating wether a spline could be successfully fitted to data}
    \item{control}{Object of class \code{grofit.control} containing list of options passed to the function as \code{control}}
    \item{fit_type}{The type of fit, here 'spline'}
}
\description{
\code{fit_growth_spline} uses \code{\link{gcFitSpline}} from the \pkg{grofit}
package to fit a smoothed spline to growth data.
}
\examples{
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_spline(df=mydata, Time, OD600)}
\dontrun{
# Fit the data given in columns Time and OD600
fit_growth_spline_(df=mydata, time_col='Time', growth_col='OD600')}
}
\seealso{
\code{\link{gcFitSpline}}
}

