% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve-package.R
\docType{package}
\name{growthcurve-package}
\alias{growthcurve-package}
\title{growthcurve: Analysis of Biological Growth}
\description{
\pkg{growthcurve} provides functions for analyzing biological growth.
}
\details{
Growth parameters such as maximum rate and yield are determined by fitting
curves to tidy data sets.

growthcurve also provides wrappers that allow tidy growth data to be analyzed
using the popular, but no-longer-maintained \pkg{grofit} package, if
installed. See \link{grofit_wrappers} for details.
}

