% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve.R
\name{growthcurve}
\alias{growthcurve}
\alias{is.growthcurve}
\title{Create \code{growthcurve} Objects}
\usage{
growthcurve(type, model, f, parameters, df, time_col, data_col)

is.growthcurve(x)
}
\arguments{
\item{type}{TODO}

\item{model}{TODO}

\item{f}{TODO}

\item{parameters}{TODO}

\item{df}{TODO}

\item{time_col}{TODO}

\item{data_col}{TODO}

\item{x}{A potential \code{growthcurve} object}
}
\value{
TODO

\code{is.growthcurve} returns a logical value indicating whether
\code{x} is a \code{growthcurve} object (\code{TRUE}) or not (\code{FALSE}).
}
\description{
\code{growthcurve} TODO

\code{is.growthcurve} determines whether or not the given object
is a growthcurve
}

