% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growthcurve.R
\docType{package}
\name{growthcurve}
\alias{growthcurve}
\alias{growthcurve-package}
\title{growthcurve: A package for the analysis of biological growth curves}
\description{
The growthcurve package provides functions for analyzing biological growth
curves. growthcurve is currently a wrapper for the \pkg{grofit} package
with a focus on incorporating growth curve analysis into modern R workflows.
These functions only work with data stored in data frames or compatible
objects. Eventually, this package will become independent of \pkg{grofit},
allowing for increased functionality.
}

