% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gcfit.R
\name{plot.gcfit}
\alias{plot.gcfit}
\title{Plot growth data and its fitted growth curve}
\usage{
\method{plot}{gcfit}(x, y = NULL, show_fit = TRUE, show_data = TRUE,
  show_maxrate = TRUE, show_asymptote = FALSE, show_lag = FALSE, ...)
}
\arguments{
\item{x}{A fit for some growth data}

\item{y}{Not used}

\item{show_fit}{Whether or not to show the fitted curve (default: \code{TRUE})}

\item{show_data}{Whether or not to show the original data (default \code{TRUE})}

\item{show_maxrate}{Whether or not to show a tangent line where the maximum growth rate occurs (default \code{TRUE})}

\item{show_asymptote}{Whether or not to indicate the maximum growth level (default \code{FALSE})}

\item{show_lag}{Whether or not to indicate where lag phase ends (default \code{FALSE}}

\item{...}{Optional formatting arguments. Includes \code{xlab}, \code{ylab},
\code{title}, \code{subtitle},
\code{fit.color}, \code{fit.linetype}, \code{fit.size},
\code{data.color}, \code{data.fill}, \code{data.shape},
\code{data.size}, \code{data.stroke},
\code{maxrate.color}, \code{maxrate.linetype}, \code{maxrate.size},
\code{asymptote.color}, \code{asymptote.linetype},
\code{asymptote.size}, \code{lag.color},
\code{lag.linetype}, \code{lag.size}}
}
\description{
Plot growth data and its fitted growth curve
}
\examples{
\dontrun{
# Get a logistic fit for some data and extract its residuals
lfit <- fit_growth_logistic(df=mydata, Time, OD600)
plot(lfit)}

}

