% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.gcfit.R
\name{plot.gcfit}
\alias{plot.gcfit}
\title{Plot growth data and its fitted growth curve}
\usage{
\method{plot}{gcfit}(x, y = NULL, ...)
}
\arguments{
\item{x}{A fit for some growth data}

\item{y}{Not used}

\item{...}{Optional arguments to plot. Note that \pkg{grofit} does not pass these along to the base graphics, so they're mostly useless at the moment.}
}
\description{
Plot growth data and its fitted growth curve
}
\examples{
\dontrun{
# Get a logistic fit for some data and extract its residuals
lfit <- fit_growth_logistic(df=mydata, Time, OD600)
plot(lfit)}
}

