% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gcfit.R
\name{plot.gcfit}
\alias{plot.gcfit}
\title{Plot growth data and its fitted growth curve}
\usage{
\method{plot}{gcfit}(x, y = NULL, show_raw = TRUE, show_maxrate = TRUE,
  show_asymptote = FALSE, ...)
}
\arguments{
\item{x}{A fit for some growth data}

\item{y}{Not used}

\item{show_raw}{Whether or not to show the original data (default \code{TRUE})}

\item{show_maxrate}{Whether or not to show a tangent line where the maximum growth rate occurs (default \code{TRUE})}

\item{show_asymptote}{Whether or not to indicate the maximum growth level (default \code{FALSE})}

\item{...}{Optional arguments to plot. Note that \pkg{grofit} does not pass these along to the base graphics, so they're mostly useless at the moment.}
}
\description{
Plot growth data and its fitted growth curve
}
\examples{
\dontrun{
# Get a logistic fit for some data and extract its residuals
lfit <- fit_growth_logistic(df=mydata, Time, OD600)
plot(lfit)}

}

