% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residuals.gcfit.R
\name{residuals.gcfit}
\alias{residuals.gcfit}
\title{Get model residuals from fits}
\usage{
\method{residuals}{gcfit}(object)
}
\arguments{
\item{object}{A fit for some growth data}
}
\value{
A numeric list containing the residuals
}
\description{
residuals gives the residuals, or the difference between the observed growth
data and the fitted values, for a given fit.
}
\examples{
\dontrun{
# Get a logistic fit for some data and extract its residuals
lfit <- fit_growth_logistic(df=mydata, Time, OD600)
residuals(lfit)}
}

